Renamed sync_user_case.py
