+++ corehq/apps/callcenter/tasks.py
@@ -5,7 +5,7 @@ from celery.task import periodic_task, task
 from celery.utils.log import get_task_logger
 
 from corehq.apps.callcenter.indicator_sets import CallCenterIndicators
-from corehq.apps.callcenter.sync_user_case import (
+from corehq.apps.callcenter.sync_usercase import (
     sync_usercases
 )
 from corehq.apps.callcenter.utils import (
