+++ corehq/apps/callcenter/tests/test_location_owners.py
@@ -2,7 +2,7 @@ from django.test import TestCase
 
 from casexml.apps.case.tests.util import delete_all_cases
 
-from corehq.apps.callcenter.sync_user_case import sync_call_center_user_case
+from corehq.apps.callcenter.sync_usercase import sync_call_center_user_case
 from corehq.apps.domain.models import CallCenterProperties
 from corehq.apps.domain.shortcuts import create_domain
 from corehq.apps.locations.models import LocationType
